remove create poprtal on tasks container and fix errors
